Abstract the underlying Spinel transport to allow light IPC
