+++ crates/api/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "ziggurat-api"
+version = "0.0.1"
+description = "Transport-agnostic JSON-RPC dispatch over the Ziggurat Zigbee stack"
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+
+[dependencies]
+ieee-802154 = { package = "ziggurat-ieee-802154", path = "../ieee-802154", version = "0.1.0" }
+spinel = { package = "ziggurat-spinel", path = "../spinel", version = "0.1.0" }
+ziggurat = { path = "../ziggurat", version = "0.0.1" }
+zigbee = { package = "ziggurat-zigbee", path = "../zigbee", version = "0.1.0" }
+
+tracing = "0.1"
+hex = "0.4.3"
+serde = { version = "1.0.219", features = ["derive"] }
+serde_json = "1.0.140"
+tokio = { version = "1.43.0", features = ["rt", "macros", "time", "sync", "io-util"] }
