+++ crates/ziggurat-server/Cargo.toml
@@ -11,12 +11,11 @@ repository.workspace = true
 [dependencies]
 spinel = { package = "ziggurat-spinel", path = "../spinel", version = "0.1.0" }
 ziggurat = { path = "../ziggurat", version = "0.0.1" }
-zigbee = { package = "ziggurat-zigbee", path = "../zigbee", version = "0.1.0" }
+ziggurat-api = { path = "../api", version = "0.0.1" }
 
 clap = { version = "4.5", features = ["derive"] }
 tracing = "0.1"
 futures-util = { version = "0.3", default-features = false, features = ["std", "sink"] }
-hex = "0.4.3"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.140"
