+++ crates/ziggurat-server/src/main.rs
@@ -2,12 +2,10 @@ use clap::{Parser, ValueEnum};
 use futures_util::{SinkExt, StreamExt};
 use serde::Deserialize;
 use serde_json::json;
-use std::sync::{Arc, Mutex};
-use std::time::Duration;
+use std::sync::Arc;
 use tokio::io::{AsyncRead, AsyncWrite};
 use tokio::net::{TcpListener, UnixListener};
 use tokio::sync::{broadcast, mpsc};
-use tokio::task::JoinHandle;
 use tokio_serial::{FlowControl, SerialPortBuilderExt};
 use tokio_tungstenite::tungstenite::Message;
 use tracing::Instrument;
@@ -15,14 +13,8 @@ use tracing_subscriber::filter::LevelFilter;
 use tracing_subscriber::prelude::*;
 use tracing_subscriber::{EnvFilter, fmt};
 
-use spinel::client::{SpinelClient, TxPriority};
-use zigbee::aps::frame::ApsDeliveryMode;
-use ziggurat::ieee_802154::types::{Eui64, Key, Nwk, PanId};
-use ziggurat::zigbee_stack::aps_security::TclkFlavor;
-use ziggurat::zigbee_stack::{
-    ApsAck, DeviceLeaveReason, NetworkBeacon, NetworkConfig, TclkSeed, Tunables,
-    WELL_KNOWN_LINK_KEY, ZigbeeNotification, ZigbeeStack,
-};
+use spinel::client::SpinelClient;
+use ziggurat_api::{Api, error_response, event, notification_to_message};
 
 const PROTOCOL_VERSION: u32 = 1;
 
@@ -30,51 +22,6 @@ const PROTOCOL_VERSION: u32 = 1;
 /// disconnected. Received frames dominate the traffic; a client that cannot keep up
 /// with them is broken.
 const OUTBOUND_QUEUE_DEPTH: usize = 1024;
-
-/// The server-level notification hub buffers this many notifications for slow
-/// connection forwarders before they start lagging.
-const NOTIFICATION_HUB_DEPTH: usize = 1024;
-
-/// The radio transmit power (in dBm) used when `configure` does not specify one.
-const DEFAULT_TX_POWER: i8 = 8;
-
-/// Big-endian colon-separated hex, the format used by zigpy for EUI64 addresses
-fn eui64_to_string(eui64: Eui64) -> String {
-    let mut bytes = eui64.to_bytes();
-    bytes.reverse();
-
-    bytes
-        .iter()
-        .map(|b| format!("{b:02x}"))
-        .collect::<Vec<_>>()
-        .join(":")
-}
-
-fn key_to_string(key: &Key) -> String {
-    key.to_bytes()
-        .iter()
-        .map(|b| format!("{b:02x}"))
-        .collect::<Vec<_>>()
-        .join(":")
-}
-
-fn network_beacon_json(beacon: &NetworkBeacon) -> serde_json::Value {
-    json!({
-        "channel": beacon.channel,
-        "source": beacon.source.map(|nwk| format!("{:04x}", nwk.0)),
-        "pan_id": format!("{:04x}", beacon.pan_id.0),
-        "extended_pan_id": eui64_to_string(beacon.extended_pan_id),
-        "permit_joining": beacon.permit_joining,
-        "stack_profile": beacon.stack_profile,
-        "protocol_version": beacon.protocol_version,
-        "router_capacity": beacon.router_capacity,
-        "end_device_capacity": beacon.end_device_capacity,
-        "device_depth": beacon.device_depth,
-        "update_id": beacon.update_id,
-        "lqi": beacon.lqi,
-        "rssi": beacon.rssi,
-    })
-}
 
 // The client wire protocol: requests carry a client-chosen correlation id; the
 // server answers each request with exactly one `response`, preceded by zero or more
@@ -88,234 +35,29 @@ struct Request {
     params: serde_json::Value,
 }
 
-fn event(id: u64, event: &str) -> serde_json::Value {
-    json!({"type": "event", "id": id, "event": event})
-}
-
-fn event_data(id: u64, event: &str, data: serde_json::Value) -> serde_json::Value {
-    json!({"type": "event", "id": id, "event": event, "data": data})
-}
-
-fn response(id: u64, result: serde_json::Value) -> serde_json::Value {
-    json!({"type": "response", "id": id, "result": result})
-}
-
-fn error_response(id: u64, code: &str, message: impl ToString) -> serde_json::Value {
-    json!({
-        "type": "response", "id": id,
-        "error": {"code": code, "message": message.to_string()},
-    })
-}
-
-fn notification(event: &str, data: serde_json::Value) -> serde_json::Value {
-    json!({"type": "notification", "event": event, "data": data})
-}
-
-// Each `params` payload deserializes into the struct matching its `method`.
-
-#[derive(Deserialize, Debug)]
-struct KeyTableEntry {
-    partner_ieee: Eui64,
-    key: Key,
-}
-
-#[derive(Deserialize, Debug)]
-struct ConfigureRequest {
-    channel: u8,
-    nwk_update_id: u8,
-    pan_id: PanId,
-    extended_pan_id: Eui64,
-    nwk_address: Nwk,
-    ieee_address: Eui64,
-    network_key: Key,
-    network_key_seq: u8,
-    network_key_tx_counter: u32,
-    tc_link_key: Option<Key>,
-    /// A TCLK seed carried over from a microcontroller stack; unique link keys are
-    /// derived from it instead of generated randomly. Requires `tclk_flavor`.
-    tclk_seed: Option<Key>,
-    tclk_flavor: Option<TclkFlavor>,
-    #[serde(default)]
-    key_table: Vec<KeyTableEntry>,
-    #[serde(default)]
-    source_routing: bool,
-    /// Radio transmit power in dBm
-    tx_power: Option<i8>,
-}
-
-#[derive(Deserialize, Debug)]
-struct SendApsRequest {
-    delivery_mode: ApsDeliveryMode,
-    /// Resolved through the address map; takes precedence over `destination`
-    destination_eui64: Option<Eui64>,
-    destination: Option<Nwk>,
-    profile_id: u16,
-    cluster_id: u16,
-    src_ep: u8,
-    dst_ep: u8,
-    aps_ack: bool,
-    aps_seq: u8,
-    radius: u8,
-    /// Hex-encoded ASDU
-    data: String,
-    /// APS-encrypt the ASDU with the destination's link key; requires a unicast
-    /// `destination_eui64`
-    #[serde(default)]
-    aps_encryption: bool,
-    #[serde(default)]
-    priority: i8,
-}
-
-#[derive(Deserialize, Debug)]
-struct EnergyScanRequest {
-    channels: Vec<u8>,
-    duration_per_channel_ms: u16,
-}
-
-#[derive(Deserialize, Debug)]
-struct NetworkScanRequest {
-    channels: Vec<u8>,
-    duration_per_channel_ms: u16,
-}
-
-#[derive(Deserialize, Debug)]
-struct PermitJoinsRequest {
-    #[serde(default)]
-    duration: u64,
-    #[serde(default = "default_accept_direct_joins")]
-    accept_direct_joins: bool,
-}
-
-const fn default_accept_direct_joins() -> bool {
-    true
-}
-
-#[derive(Deserialize, Debug)]
-struct SetProvisionalKeyRequest {
-    ieee: Eui64,
-    key: Key,
-}
-
-#[derive(Deserialize, Debug)]
-struct SetChannelRequest {
-    channel: u8,
-}
-
-#[derive(Deserialize, Debug)]
-struct SetNwkUpdateIdRequest {
-    nwk_update_id: u8,
-}
-
-fn notification_to_message(notification_event: ZigbeeNotification) -> serde_json::Value {
-    match notification_event {
-        ZigbeeNotification::ReceivedApsCommand {
-            source,
-            destination,
-            group,
-            profile_id,
-            cluster_id,
-            src_ep,
-            dst_ep,
-            lqi,
-            rssi,
-            data,
-        } => notification(
-            "received_aps_command",
-            json!({
-                "source": hex::encode(source.to_bytes()),
-                "destination": hex::encode(destination.to_bytes()),
-                "group": group,
-                "profile_id": profile_id,
-                "cluster_id": cluster_id, "src_ep": src_ep, "dst_ep": dst_ep,
-                "lqi": lqi, "rssi": rssi, "data": hex::encode(data),
-            }),
-        ),
-        ZigbeeNotification::FrameCounterUpdate { frame_counter } => notification(
-            "frame_counter_update",
-            json!({"frame_counter": frame_counter}),
-        ),
-        ZigbeeNotification::LinkKeyUpdate { ieee, key } => notification(
-            "link_key_update",
-            json!({
-                "ieee": eui64_to_string(ieee),
-                "key": key_to_string(&key),
-            }),
-        ),
-        ZigbeeNotification::DeviceJoined { nwk, ieee, parent } => notification(
-            "device_joined",
-            json!({
-                "nwk": hex::encode(nwk.to_bytes()),
-                "ieee": eui64_to_string(ieee),
-                "parent": hex::encode(parent.to_bytes()),
-            }),
-        ),
-        ZigbeeNotification::DeviceLeft { nwk, ieee, reason } => {
-            let mut params = json!({
-                "nwk": hex::encode(nwk.to_bytes()),
-                "ieee": ieee.map(eui64_to_string),
-            });
-            match reason {
-                DeviceLeaveReason::Announced { rejoin } => {
-                    params["reason"] = json!("announced");
-                    params["rejoin"] = json!(rejoin);
-                }
-                DeviceLeaveReason::RouterReported {
-                    router,
-                    router_ieee,
-                } => {
-                    params["reason"] = json!("router_reported");
-                    params["router"] = json!(hex::encode(router.to_bytes()));
-                    params["router_ieee"] = json!(router_ieee.map(eui64_to_string));
-                }
-                DeviceLeaveReason::KeepaliveTimeout => {
-                    params["reason"] = json!("keepalive_timeout");
-                }
-            }
-            notification("device_left", params)
-        }
-        ZigbeeNotification::ApsDecryptionFailure {
-            source,
-            source_ieee,
-            frame_counter,
-            key_id,
-        } => notification(
-            "aps_decryption_failure",
-            json!({
-                "source": hex::encode(source.to_bytes()),
-                "source_ieee": eui64_to_string(source_ieee),
-                "frame_counter": frame_counter,
-                "key_id": key_id,
-            }),
-        ),
-    }
-}
-
 pub struct ZigguratServer {
-    serial: SerialConfig,
-    /// The Spinel client owns the serial port for the lifetime of the process: it is
-    /// opened lazily by the first command that needs it and never reopened, so stack
-    /// replacement cannot race a straggling port handle (`EBUSY`)
-    spinel: Mutex<Option<Arc<SpinelClient>>>,
-    stack: Mutex<Option<Arc<ZigbeeStack>>>,
-    /// The server-level notification hub: connections subscribe to it, and it
-    /// survives stack replacement (the forwarder task is swapped instead)
-    notification_tx: broadcast::Sender<ZigbeeNotification>,
-    notification_forwarder: Mutex<Option<JoinHandle<()>>>,
+    api: Arc<Api>,
 }
 
 impl ZigguratServer {
     /// The serial port is not opened and the Zigbee stack is not created until a
     /// client sends a command that needs them.
     pub fn new(serial: SerialConfig) -> Self {
-        let (notification_tx, _) = broadcast::channel(NOTIFICATION_HUB_DEPTH);
+        // The factory opens the serial port on first use, building the process-lifetime
+        // Spinel client. Without flow control the RCP's UART drops bytes under load,
+        // corrupting host->RCP frames ("Framing error" + command timeout).
+        let api = Api::new(Box::new(move || {
+            let port = tokio_serial::new(&serial.device, serial.baudrate)
+                .flow_control(serial.flow_control.into())
+                .open_native_async()
+                .map_err(|e| e.to_string())?;
 
-        Self {
-            serial,
-            spinel: Mutex::new(None),
-            stack: Mutex::new(None),
-            notification_tx,
-            notification_forwarder: Mutex::new(None),
-        }
+            let client = Arc::new(SpinelClient::new(port));
+            client.spawn_reader();
+            Ok(client)
+        }));
+
+        Self { api }
     }
 
     pub async fn run(self: Arc<Self>, listen_addr: &str) -> std::io::Result<()> {
@@ -369,32 +111,6 @@ impl ZigguratServer {
         });
     }
 
-    fn current_stack(&self) -> Option<Arc<ZigbeeStack>> {
-        self.stack.lock().unwrap().clone()
-    }
-
-    /// The process-lifetime Spinel client, opening the serial port on first use.
-    fn spinel_client(&self) -> Result<Arc<SpinelClient>, tokio_serial::Error> {
-        let mut spinel = self.spinel.lock().unwrap();
-
-        if let Some(spinel) = &*spinel {
-            return Ok(spinel.clone());
-        }
-
-        // Without flow control the RCP's UART drops bytes under load, corrupting
-        // host->RCP frames ("Framing error" + command timeout)
-        let port = tokio_serial::new(&self.serial.device, self.serial.baudrate)
-            .flow_control(self.serial.flow_control.into())
-            .open_native_async()?;
-
-        let client = Arc::new(SpinelClient::new(port));
-        client.spawn_reader();
-        *spinel = Some(client.clone());
-        drop(spinel);
-
-        Ok(client)
-    }
-
     async fn handle_connection<S>(
         self: &Arc<Self>,
         socket: S,
@@ -423,7 +139,7 @@ impl ZigguratServer {
             let _ = sink.close().await;
         });
 
-        let state = if self.current_stack().is_some() {
+        let state = if self.api.is_configured() {
             "running"
         } else {
             "awaiting_configuration"
@@ -433,7 +149,7 @@ impl ZigguratServer {
             .await?;
 
         // Forward hub notifications to this connection
-        let mut notification_rx = self.notification_tx.subscribe();
+        let mut notification_rx = self.api.subscribe();
         let notification_outbound = outbound_tx.clone();
         let forwarder_addr = addr.to_owned();
         let notification_forwarder = tokio::spawn(async move {
@@ -491,7 +207,7 @@ impl ZigguratServer {
     /// Dispatches a request, spawning everything that can block on network activity:
     /// a command waiting on a slow device must never delay other commands.
     fn dispatch(self: &Arc<Self>, request: Request, outbound: mpsc::Sender<serde_json::Value>) {
-        let server = self.clone();
+        let api = self.api.clone();
 
         // One span per request so the handler work nests under it and the close line
         // reports the full request-to-response latency.
@@ -501,447 +217,12 @@ impl ZigguratServer {
             async move {
                 let Request { id, method, params } = request;
 
-                let message = match method.as_str() {
-                    "ping" => server.handle_ping(id).await,
-                    "configure" => server.handle_configure(id, params).await,
-                    "get_hw_address" => server.handle_get_hw_address(id).await,
-                    "get_network_info" => server.handle_get_network_info(id),
-                    "send_aps" => server.handle_send_aps(id, params, &outbound).await,
-                    "energy_scan" => server.handle_energy_scan(id, params, &outbound).await,
-                    "network_scan" => server.handle_network_scan(id, params, &outbound).await,
-                    "permit_joins" => server.handle_permit_joins(id, params),
-                    "set_provisional_key" => server.handle_set_provisional_key(id, params),
-                    "set_nwk_update_id" => server.handle_set_nwk_update_id(id, params),
-                    "set_channel" => server.handle_set_channel(id, params).await,
-                    _ => error_response(id, "unknown_method", method),
-                };
+                let message = api.dispatch(id, &method, params, &outbound).await;
 
                 let _ = outbound.send(message).await;
             }
             .instrument(span),
         );
-    }
-
-    /// Liveness probe. Yielding makes the reply round-trip through the runtime like
-    /// every real command, so a starved executor shows up in the latency.
-    async fn handle_ping(&self, id: u64) -> serde_json::Value {
-        tokio::task::yield_now().await;
-
-        response(id, json!({"status": "pong"}))
-    }
-
-    /// (Re)initializes the Zigbee stack. The stack deliberately outlives client
-    /// connections; reconfiguring replaces it wholesale.
-    #[allow(clippy::significant_drop_tightening)]
-    async fn handle_configure(&self, id: u64, params: serde_json::Value) -> serde_json::Value {
-        let request: ConfigureRequest = match serde_json::from_value(params) {
-            Ok(request) => request,
-            Err(e) => return error_response(id, "invalid_request", e),
-        };
-
-        let tclk_seed = match (request.tclk_seed, request.tclk_flavor) {
-            (Some(seed), Some(flavor)) => Some(TclkSeed { seed, flavor }),
-            (None, None) => None,
-            _ => {
-                return error_response(
-                    id,
-                    "invalid_request",
-                    "tclk_seed and tclk_flavor must be provided together",
-                );
-            }
-        };
-
-        // A replaced stack must be fully stopped before its successor registers its
-        // own receivers with the shared Spinel client
-        let old_stack = self.stack.lock().unwrap().take();
-        if let Some(old_stack) = old_stack {
-            tracing::info!("Replacing the running Zigbee stack");
-            old_stack.shutdown().await;
-        }
-
-        let old_forwarder = self.notification_forwarder.lock().unwrap().take();
-        if let Some(old_forwarder) = old_forwarder {
-            old_forwarder.abort();
-        }
-
-        tracing::info!("Initializing Zigbee stack with new settings...");
-        let spinel = match self.spinel_client() {
-            Ok(s) => s,
-            Err(e) => return error_response(id, "serial_port_error", e),
-        };
-
-        let (stack, mut stack_notification_rx) = ZigbeeStack::new(
-            spinel,
-            NetworkConfig {
-                channel: request.channel,
-                update_id: request.nwk_update_id,
-                pan_id: request.pan_id,
-                extended_pan_id: request.extended_pan_id,
-                network_address: request.nwk_address,
-                ieee_address: request.ieee_address,
-                network_key: request.network_key,
-                network_key_seq_number: request.network_key_seq,
-                network_key_tx_counter: request.network_key_tx_counter,
-                tc_link_key: request.tc_link_key.unwrap_or(WELL_KNOWN_LINK_KEY),
-                tclk_seed,
-                tx_power: request.tx_power.unwrap_or(DEFAULT_TX_POWER),
-                source_routing: request.source_routing,
-            },
-            Tunables::new(),
-        );
-
-        // Restore unique trust center link keys negotiated in earlier sessions
-        if !request.key_table.is_empty() {
-            let mut core = stack.state.core.lock();
-
-            for entry in request.key_table {
-                core.aib
-                    .aps_security
-                    .restore_device_key(entry.partner_ieee, entry.key);
-            }
-
-            tracing::info!(
-                "Restored {} trust center link keys",
-                core.aib.aps_security.device_key_count()
-            );
-        }
-
-        // The success response is the client's permission to send commands: the
-        // network must be fully up (RCP reset handled, radio programmed) before
-        // replying, or the client's first command would race with the boot-time reset.
-        if let Err(e) = stack.start_network().await {
-            stack.shutdown().await;
-            return error_response(id, "network_start_failed", e);
-        }
-
-        let stack_clone = stack.clone();
-        stack.spawn_tracked(async move {
-            stack_clone.run().await;
-        });
-
-        // Pump the stack's notifications into the server-level hub
-        let hub_tx = self.notification_tx.clone();
-        let forwarder = tokio::spawn(async move {
-            while let Ok(event) = stack_notification_rx.recv().await {
-                // Send errors just mean no client is connected right now
-                let _ = hub_tx.send(event);
-            }
-        });
-
-        *self.stack.lock().unwrap() = Some(stack);
-        *self.notification_forwarder.lock().unwrap() = Some(forwarder);
-
-        tracing::info!("Zigbee stack initialized and running.");
-        response(id, json!({"status": "success"}))
-    }
-
-    /// Updates the `nwkUpdateId` advertised in beacons, the companion to
-    /// `set_channel` during a network-wide channel migration.
-    fn handle_set_nwk_update_id(&self, id: u64, params: serde_json::Value) -> serde_json::Value {
-        let request: SetNwkUpdateIdRequest = match serde_json::from_value(params) {
-            Ok(request) => request,
-            Err(e) => return error_response(id, "invalid_request", e),
-        };
-
-        let Some(stack) = self.current_stack() else {
-            return error_response(id, "not_configured", "no stack is running");
-        };
-
-        stack.set_nwk_update_id(request.nwk_update_id);
-        response(id, json!({"status": "success"}))
-    }
-
-    /// Retunes the radio to a new channel, the coordinator's half of a network-wide
-    /// channel migration; broadcasting `Mgmt_NWK_Update_req` to the other devices is
-    /// the client's job.
-    async fn handle_set_channel(&self, id: u64, params: serde_json::Value) -> serde_json::Value {
-        let request: SetChannelRequest = match serde_json::from_value(params) {
-            Ok(request) => request,
-            Err(e) => return error_response(id, "invalid_request", e),
-        };
-
-        let Some(stack) = self.current_stack() else {
-            return error_response(id, "not_configured", "no stack is running");
-        };
-
-        match stack.set_channel(request.channel).await {
-            Ok(()) => response(id, json!({"status": "success"})),
-            Err(e) => error_response(id, "set_channel_failed", e),
-        }
-    }
-
-    /// Reads back the running network's settings, the counterpart of `configure`.
-    /// While the stack runs, the server is the authoritative holder of the live state
-    /// (e.g. frame counters), not the client that configured it.
-    #[allow(clippy::significant_drop_tightening)]
-    fn handle_get_network_info(&self, id: u64) -> serde_json::Value {
-        let Some(stack) = self.current_stack() else {
-            return error_response(id, "not_configured", "no stack is running");
-        };
-
-        let state = &stack.state;
-        let core = state.core.lock();
-        let nwk_security = &core.nib.nwk_security;
-        let aps_security = &core.aib.aps_security;
-        let tclk_seed = &stack.config.tclk_seed;
-
-        response(
-            id,
-            json!({
-                "channel": core.mac.channel,
-                "nwk_update_id": core.nib.update_id,
-                "pan_id": format!("{:04x}", core.mac.pan_id.0),
-                "extended_pan_id": eui64_to_string(state.extended_pan_id),
-                "nwk_address": format!("{:04x}", state.network_address.as_u16()),
-                "ieee_address": eui64_to_string(state.ieee_address),
-                "network_key": key_to_string(&nwk_security.network_key()),
-                "network_key_seq": nwk_security.key_seq_number(),
-                "network_key_tx_counter": nwk_security.outgoing_frame_counter(),
-                "tc_link_key": key_to_string(&stack.config.tc_link_key),
-                "tx_power": stack.config.tx_power,
-                "tclk_seed": tclk_seed.as_ref().map(|tclk| hex::encode(tclk.seed.to_bytes())),
-                "tclk_flavor": tclk_seed.as_ref().map(|tclk| match tclk.flavor {
-                    TclkFlavor::ZStack => "zstack",
-                    TclkFlavor::Ezsp => "ezsp",
-                }),
-                "key_table": aps_security
-                    .device_keys()
-                    .map(|(partner_ieee, entry)| json!({
-                        "partner_ieee": eui64_to_string(partner_ieee),
-                        "key": key_to_string(&entry.key),
-                    }))
-                    .collect::<Vec<_>>(),
-            }),
-        )
-    }
-
-    /// Reads the radio's factory-programmed EUI64, which a client needs before it can
-    /// form a network with `configure`.
-    async fn handle_get_hw_address(&self, id: u64) -> serde_json::Value {
-        let spinel = match self.spinel_client() {
-            Ok(s) => s,
-            Err(e) => return error_response(id, "serial_port_error", e),
-        };
-
-        match spinel.get_hw_address().await {
-            Ok(ieee) => response(id, json!({"ieee_address": eui64_to_string(ieee)})),
-            Err(e) => error_response(id, "hw_address_failed", e),
-        }
-    }
-
-    async fn handle_send_aps(
-        &self,
-        id: u64,
-        params: serde_json::Value,
-        outbound: &mpsc::Sender<serde_json::Value>,
-    ) -> serde_json::Value {
-        let request: SendApsRequest = match serde_json::from_value(params) {
-            Ok(request) => request,
-            Err(e) => return error_response(id, "invalid_request", e),
-        };
-
-        let Some(stack) = self.current_stack() else {
-            return error_response(id, "not_configured", "no stack is running");
-        };
-
-        // A network address is authoritative when given (`destination_eui64` then only
-        // selects the link key); EUI64-only packets are resolved through the address map
-        let destination = match (request.destination_eui64, request.destination) {
-            (_, Some(nwk)) => nwk,
-            (Some(eui64), None) => {
-                let nwk = stack.state.core.lock().nib.address_map.nwk_for(eui64);
-
-                match nwk {
-                    Some(nwk) => nwk,
-                    None => {
-                        return error_response(
-                            id,
-                            "unknown_destination_eui64",
-                            format!("{eui64:?}"),
-                        );
-                    }
-                }
-            }
-            (None, None) => {
-                return error_response(id, "missing_destination", "no destination given");
-            }
-        };
-
-        let asdu = match hex::decode(&request.data) {
-            Ok(asdu) => asdu,
-            Err(e) => return error_response(id, "invalid_data", e),
-        };
-
-        // Link keys are pairwise: encryption needs a unicast EUI64-addressed target
-        let aps_security = if request.aps_encryption {
-            match (request.destination_eui64, request.delivery_mode) {
-                (Some(eui64), ApsDeliveryMode::Unicast) => Some(eui64),
-                _ => {
-                    return error_response(
-                        id,
-                        "invalid_request",
-                        "aps_encryption requires a unicast destination_eui64",
-                    );
-                }
-            }
-        } else {
-            None
-        };
-
-        let ack_waiter = match stack
-            .send_aps_command(
-                request.delivery_mode,
-                destination,
-                request.profile_id,
-                request.cluster_id,
-                request.src_ep,
-                request.dst_ep,
-                if request.aps_ack {
-                    ApsAck::Request
-                } else {
-                    ApsAck::None
-                },
-                request.radius,
-                request.aps_seq,
-                asdu,
-                aps_security,
-                TxPriority(request.priority),
-            )
-            .await
-        {
-            Ok(ack_waiter) => ack_waiter,
-            Err(e) => return error_response(id, "transmit_failed", e),
-        };
-
-        // The frame is on the air (or extracted from the indirect queue); the
-        // terminal response then reports end-to-end delivery when an ack was requested
-        let _ = outbound.send(event(id, "transmitted")).await;
-
-        match ack_waiter {
-            None => response(id, json!({"status": "sent"})),
-            Some(waiter) => match stack.wait_aps_ack(waiter).await {
-                Ok(()) => response(id, json!({"status": "delivered"})),
-                Err(e) => error_response(id, "aps_ack_timeout", e),
-            },
-        }
-    }
-
-    async fn handle_energy_scan(
-        &self,
-        id: u64,
-        params: serde_json::Value,
-        outbound: &mpsc::Sender<serde_json::Value>,
-    ) -> serde_json::Value {
-        let request: EnergyScanRequest = match serde_json::from_value(params) {
-            Ok(request) => request,
-            Err(e) => return error_response(id, "invalid_request", e),
-        };
-
-        let Some(stack) = self.current_stack() else {
-            return error_response(id, "not_configured", "no stack is running");
-        };
-
-        let (result_tx, mut result_rx) = mpsc::channel::<(u8, i8)>(32);
-
-        // The scan runs on its own task so it always reaches its channel restore, even if
-        // this request's task is dropped. Its only sender lives until the scan ends, so
-        // the drain loop below terminates exactly when the scan is done.
-        let duration = Duration::from_millis(u64::from(request.duration_per_channel_ms));
-        let scan = tokio::spawn(async move {
-            stack
-                .energy_scan(&request.channels, duration, result_tx)
-                .await
-        });
-
-        while let Some((channel, rssi)) = result_rx.recv().await {
-            let _ = outbound
-                .send(event_data(
-                    id,
-                    "energy_result",
-                    json!({"channel": channel, "rssi": rssi}),
-                ))
-                .await;
-        }
-
-        match scan.await {
-            Ok(Ok(())) => response(id, json!({"status": "complete"})),
-            Ok(Err(e)) => error_response(id, "energy_scan_failed", e),
-            Err(e) => error_response(id, "energy_scan_failed", e),
-        }
-    }
-
-    async fn handle_network_scan(
-        &self,
-        id: u64,
-        params: serde_json::Value,
-        outbound: &mpsc::Sender<serde_json::Value>,
-    ) -> serde_json::Value {
-        let request: NetworkScanRequest = match serde_json::from_value(params) {
-            Ok(request) => request,
-            Err(e) => return error_response(id, "invalid_request", e),
-        };
-
-        let Some(stack) = self.current_stack() else {
-            return error_response(id, "not_configured", "no stack is running");
-        };
-
-        let (found_tx, mut found_rx) = mpsc::channel::<NetworkBeacon>(32);
-
-        // The scan runs on its own task so it always reaches its channel restore, even if
-        // this request's task is dropped. Its only sender lives until the scan ends, so
-        // the drain loop below terminates exactly when the scan is done.
-        let duration = Duration::from_millis(u64::from(request.duration_per_channel_ms));
-        let scan = tokio::spawn(async move {
-            stack
-                .network_scan(&request.channels, duration, found_tx)
-                .await
-        });
-
-        while let Some(beacon) = found_rx.recv().await {
-            let _ = outbound
-                .send(event_data(
-                    id,
-                    "network_found",
-                    network_beacon_json(&beacon),
-                ))
-                .await;
-        }
-
-        match scan.await {
-            Ok(Ok(())) => response(id, json!({"status": "complete"})),
-            Ok(Err(e)) => error_response(id, "network_scan_failed", e),
-            Err(e) => error_response(id, "network_scan_failed", e),
-        }
-    }
-
-    fn handle_permit_joins(&self, id: u64, params: serde_json::Value) -> serde_json::Value {
-        let request: PermitJoinsRequest = match serde_json::from_value(params) {
-            Ok(request) => request,
-            Err(e) => return error_response(id, "invalid_request", e),
-        };
-
-        let Some(stack) = self.current_stack() else {
-            return error_response(id, "not_configured", "no stack is running");
-        };
-
-        stack.permit_joins(request.duration, request.accept_direct_joins);
-
-        response(id, json!({"status": "success"}))
-    }
-
-    fn handle_set_provisional_key(&self, id: u64, params: serde_json::Value) -> serde_json::Value {
-        let request: SetProvisionalKeyRequest = match serde_json::from_value(params) {
-            Ok(request) => request,
-            Err(e) => return error_response(id, "invalid_request", e),
-        };
-
-        let Some(stack) = self.current_stack() else {
-            return error_response(id, "not_configured", "no stack is running");
-        };
-
-        stack.set_provisional_key(request.ieee, request.key);
-
-        response(id, json!({"status": "success"}))
     }
 }
 
@@ -962,7 +243,7 @@ impl From<FlowControlMode> for FlowControl {
     }
 }
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct SerialConfig {
     device: String,
     baudrate: u32,
